+++ assets/javascripts/kitten/components/molecules/alert/index.js
@@ -1,7 +1,8 @@
 import React, { useEffect, useRef, useState } from 'react'
 import PropTypes from 'prop-types'
 import styled, { keyframes } from 'styled-components'
-import { CloseButton } from '../../../components/molecules/buttons/close-button'
+import { CrossIconNext } from '../../../components/graphics/icons-next/cross-icon-next'
+import { IconBadge } from '../../../components/atoms/icon-badge'
 import COLORS from '../../../constants/colors-config'
 import TYPOGRAPHY from '../../../constants/typography-config'
 import { ScreenConfig } from '../../../constants/screen-config'
@@ -15,12 +16,18 @@ const fadeOut = keyframes`
 
 const AlertWrapper = styled.div`
   ${TYPOGRAPHY.fontStyles.light};
+  margin: ${pxToRem(10)};
+  border-radius: ${pxToRem(8)};
   position: relative;
   overflow: hidden;
   background-color: ${COLORS.primary5};
-  color: ${COLORS.primary1};
+  color: ${COLORS.font1};
   display: flex;
   align-items: flex-start;
+
+  [href] {
+    color: ${COLORS.primary1};
+  }
 
   .k-Alert__text {
     padding: ${pxToRem(13)} ${pxToRem(20)};
@@ -28,17 +35,33 @@ const AlertWrapper = styled.div`
     font-size: ${stepToRem(-1)};
 
     @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
-      text-align: center;
+      display: flex;
+      justify-content: center;
     }
   }
 
-  .k-Alert__button {
+  .k-Alert__iconBadge {
+    margin-right: ${pxToRem(20)};
+    background-color: ${COLORS.primary1};
+    border-color: ${COLORS.primary4};
+  }
+
+  button {
+    display: flex;
     flex: 0 0 auto;
     transition: all 0.2s ease;
-
+    align-self: stretch;
+    align-items: center;
+    padding-right: ${pxToRem(20)};
+    
     svg,
     svg path {
       transition: fill 0.2s ease;
+      fill: ${COLORS.primary1};
+      
+      :hover {
+        fill: ${COLORS.primary3};
+      }
     }
 
     &:focus {
@@ -60,18 +83,75 @@ const AlertWrapper = styled.div`
   }
 
   &.k-Alert--success {
-    color: ${COLORS.valid};
     background-color: ${COLORS.tertiary1};
+    
+    [href] {
+      color: ${COLORS.valid};
+    }
+  
+    .k-Alert__iconBadge {
+      background-color: ${COLORS.valid};
+      border-color: ${COLORS.tertiary2};
+    }
+    
+    button {
+      svg,
+      svg path {
+        fill: ${COLORS.valid};
+
+        :hover {
+          fill: ${COLORS.tertiary2};
+        }
+      }
+    }
   }
 
   &.k-Alert--error {
-    color: ${COLORS.error};
     background-color: ${COLORS.error2};
+
+    [href] {
+      color: ${COLORS.error};
+    }
+  
+    .k-Alert__iconBadge {
+      background-color: ${COLORS.error};
+      border-color: ${COLORS.error3};
+    }
+
+    button {
+      svg,
+      svg path {
+        fill: ${COLORS.error};
+
+        :hover {
+          fill: ${COLORS.error4};
+        }
+      }
+    }
   }
 
   &.k-Alert--warning {
-    color: ${COLORS.warning};
     background-color: ${COLORS.warning2};
+
+    [href] {
+      color: ${COLORS.orange3};
+    }
+  
+    .k-Alert__iconBadge {
+      background-color: ${COLORS.orange3};
+      border-color: ${COLORS.orange};
+    }
+  
+    button {
+      svg,
+      svg path {
+        fill: ${COLORS.warning};
+
+        :hover {
+          fill: "#A47600";
+        }
+      }
+    }
   }
 
   &.k-Alert--hasCloseButton {
@@ -98,6 +178,8 @@ export const Alert = ({
   closeButtonLabel,
   children,
   onAfterClose,
+  icon,
+  iconBadgeBorderColor,
   ...others
 }) => {
   const [isTrashed, trashIt] = useState(false)
@@ -131,15 +213,31 @@ export const Alert = ({
       {...others}
     >
       <>
-        <div className="k-Alert__text">{children}</div>
+        <div className="k-Alert__text">
+          {icon && (
+            <IconBadge
+              className="k-Alert__iconBadge"
+              size="tiny"
+              children={icon}
+              border={{
+                width: 2,
+                color: iconBadgeBorderColor,
+                style: 'solid',
+              }}
+            />
+          )}
+          {children}
+        </div>
 
         {closeButton && (
-          <CloseButton
-            modifier="carbon"
-            closeButtonLabel={closeButtonLabel}
-            className="k-Alert__button"
-            onClick={() => setMounted(false)}
-          />
+          <button className="k-u-reset-button">
+            <CrossIconNext
+              onClick={() => setMounted(false)}
+              title={closeButtonLabel}
+              width={12}
+              height={12}
+            />
+          </button>
         )}
       </>
     </AlertWrapper>
@@ -154,6 +252,8 @@ Alert.propTypes = {
   closeButton: PropTypes.bool,
   closeButtonLabel: PropTypes.string,
   onAfterClose: PropTypes.func,
+  icon: PropTypes.node,
+  iconBadgeBorderColor: PropTypes.string,
 }
 
 Alert.defaultProps = {
@@ -164,4 +264,6 @@ Alert.defaultProps = {
   closeButton: false,
   closeButtonLabel: 'Close',
   onAfterClose: () => {},
+  icon: '',
+  iconBadgeBorderColor: COLORS.primary4,
 }